Add resume button and modular settings
